validando o post e delete de Produto
fazendo algumas verificações para post e delete

+++ src/main/java/com/farmacia/domain/service/ProdutoService.java
@@ -1,5 +1,6 @@
 package com.farmacia.domain.service;
 
+import java.util.ArrayList;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,7 +16,9 @@ public class ProdutoService {
 	private ProdutoRepository produtoRepositorio;
 
 	public ProdutoEntidade save(ProdutoEntidade produto) {
-		return produtoRepositorio.save(produto);
+		ArrayList<ProdutoEntidade> busca= (ArrayList<ProdutoEntidade>) produtoRepositorio.findByNomeContainingIgnoreCase(produto.getNome());
+		if (busca.size()> 0) return null; //já existe um produto com esse nome
+		else return produtoRepositorio.save(produto);
 	}
 
 	public ProdutoEntidade alterProduto(long id_produto, ProdutoEntidade produto) {
@@ -29,8 +32,12 @@ public class ProdutoService {
 		}
 	}
 	
-	public void deleteProduto(long id_produto) {
-		produtoRepositorio.deleteById(id_produto);
+	public boolean deleteProduto(long id_produto) {
+		if (produtoRepositorio.existsById(id_produto)) {
+			//existe um produto com o id passado
+			produtoRepositorio.deleteById(id_produto);
+			return true;
+		}else return false;
 	}
 	
 	public ProdutoRepository getProdutoRepositorio() {
